#625, #365, code upgrade, security, errors, fxes

+++ myhoard/apps/collections/items/models.py
@@ -1,13 +1,16 @@
+import logging
 from datetime import datetime
+
+from werkzeug.exceptions import Forbidden, NotFound
 
 from flask import g
 from flask.ext.mongoengine import Document
-from mongoengine import StringField, ListField, ObjectIdField, DateTimeField, \
-    PointField
+from mongoengine import StringField, ListField, ObjectIdField, DateTimeField, PointField
 
 from myhoard.apps.media.models import Media
-from myhoard.apps.common.utils import make_order_by_for_query, \
-    make_item_search_query
+from myhoard.apps.common.utils import make_order_by_for_query, make_item_search_query
+
+logger = logging.getLogger(__name__)
 
 
 class Item(Document):
@@ -20,8 +23,29 @@ class Item(Document):
     collection = ObjectIdField(required=True)
     owner = ObjectIdField()
 
-    def __repr__(self):
-        return '<Item {}>'.format(self.name)
+    @property
+    def public(self):
+        from myhoard.apps.collections.models import Collection
+
+        return bool(Collection.objects(id=self.collection, public=True).count())
+
+    def __str__(self):
+        return '<{} {}>'.format(type(self).__name__, self.id)
+
+    @classmethod
+    def get_visible_or_404(cls, item_id):
+        from myhoard.apps.collections.models import Collection
+
+        item = cls.objects.get_or_404(id=item_id)
+        Collection.get_visible_or_404(item.collection)
+
+        logger.debug('get_visible_or_404 dump:\nitem: {}'.format(item._data))
+
+        return item
+
+    @classmethod
+    def get_all(cls, params, collection_id):
+        return cls.objects(make_item_search_query(params, collection_id)).order_by(*make_order_by_for_query(params))
 
     @classmethod
     def create(cls, **kwargs):
@@ -33,26 +57,34 @@ class Item(Document):
 
         if 'location' in item:
             item.location = {
-                "type": "Point",
-                "coordinates": [item.location.get('lng'),
-                                item.location.get('lat')]
+                'type': 'Point',
+                'coordinates': [item.location.get('lng'), item.location.get('lat')]
             }
 
         item.save()
+        logger.info('{} created'.format(item))
+
+        logger.info('Updating {} Media'.format(item))
         Media.create_from_item(item)
 
         return item
 
     @classmethod
     def put(cls, item_id, **kwargs):
-        item = cls.objects.get_or_404(id=item_id, owner=g.user)
+        item = cls.objects.get_or_404(id=item_id)
+        if item.owner != g.user:
+            raise Forbidden('Only collection owner can edit items') if item.public else NotFound()
+
         update_item = cls(**kwargs)
 
         return cls.update(item, update_item)
 
     @classmethod
     def patch(cls, item_id, **kwargs):
-        item = cls.objects.get_or_404(id=item_id, owner=g.user)
+        item = cls.objects.get_or_404(id=item_id)
+        if item.owner != g.user:
+            raise Forbidden('Only collection owner can edit items') if item.public else NotFound()
+
         update_item = cls()
 
         for field in item._fields:
@@ -69,34 +101,28 @@ class Item(Document):
 
         if 'location' in update_item:
             update_item.location = {
-                "type": "Point",
-                "coordinates": [update_item.location.get('lng'),
-                                update_item.location.get('lat')]
+                'type': 'Point',
+                'coordinates': [update_item.location.get('lng'), update_item.location.get('lat')]
             }
 
         update_item.save()
-        Media.update_from_item(item, update_item)
+        logger.info('{} updated'.format(item))
 
         return update_item
 
     @classmethod
     def delete(cls, item_id):
-        item = cls.objects.get_or_404(id=item_id, owner=g.user)
+        item = cls.objects.get_or_404(id=item_id)
+        if item.owner != g.user:
+            raise Forbidden('Only collection owner can delete items') if item.public else NotFound()
 
-        super(cls, item).delete()
+        logger.info('Deleting {} Media'.format(item))
         Media.delete_from_item(item)
 
-        return item
+        super(cls, item).delete()
+        logger.info('{} deleted'.format(item))
 
     @classmethod
     def delete_from_collection(cls, collection):
-        for item in cls.objects(collection=collection.id):
-            Media.delete_from_item(item)
-
-            item.delete()
-
-    @classmethod
-    def get_all(cls, params, collection_id):
-        return cls.objects(
-            make_item_search_query(params, collection_id)).order_by(
-            *make_order_by_for_query(params))
+        for item_id in cls.objects(collection=collection.id).scalar('id'):
+            cls.delete(item_id)
